requirements.txt
async-timeout==5.0.1
certifi==2025.4.26
charset-normalizer==3.4.2
greenlet==3.2.3
idna==3.10
numpy==2.0.2
pandas==2.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==6.2.0
requests==2.32.4
six==1.17.0
spotipy==2.25.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
